Closure passed to `find_chunk_edge_cond` is now FnMut
There's no reason to limit this to `Fn`, and `FnMut` may be useful: it
would be possible to perform a min/max size check by mutating a counter,
for example.

This is technically a breaking change: Any implementations outside this
crate may have overridden the `find_chunk_edge_cond` function, and they
would no longer have the correct trait bounds on the implementation.